backend/requirements.txt
fastapi[standard]
uvicorn
python-dotenv==1.0.0
pymongo==4.5.0
groq
requests
motor